editing: extract MapEditService from ObjectCommandController
Continues the controller split: the map-wide operations — newEmptyMap,
clearElevationObjects, copyElevation — move into MapEditService, built on the
shared UndoBatcher and borrowing ScriptEditService for the script
snapshot/restore/erase that clearing an elevation needs. ObjectCommandController
keeps the public methods as thin delegations. 235 tests pass.
